requirement.txt
langchain
streamlit
asyncio
python-dotenv